requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.18
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
prettytable==3.6.0
pyasn1==0.4.8
pydantic==1.10.5
pytest==7.2.1
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
virtualenv==20.19.0
wcwidth==0.2.6
